Smite the abomination
Dec. 9th, 2024. 

Ruby is reverting some fake rtc business, and in so doing accidentally removes zone identifiers from the gitignore.

Little do they know the tragedy that awaits them.

Dec. 13th, 2024.

Doodle's computer, stealthy and wise in the ways of the Zone Identifier, decides the time is ripe to slip in an unsuspecting GH Desktop explosion while the shields are down. The commit is made, the damage is done.

Dec. 13th, 2024. (4 commits later)

Doodle notices that zone identifiers have been removed from the gitignore, and dutifully adds them back in.

It is too late.

They are already here.